database check for migrate script before post
